Global State Repository Pattern implementiert
- Für Global State das Repository Pattern implementiert.
