+++ lib/models/global_state/global_state_interface.dart
@@ -1,0 +1,5 @@
+abstract class GlobalStateInterface {
+  void create();
+  Future<int> getBookingSerieIndex();
+  void increaseBookingSerieIndex();
+}
