+++ lib/models/global_state/global_state_repository.dart
@@ -1,43 +1,29 @@
 import 'package:hive/hive.dart';
 
 import '/utils/consts/hive_consts.dart';
+import 'global_state_interface.dart';
+import 'global_state_model.dart';
 
-@HiveType(typeId: globalStateTypeId)
-class GlobalState extends HiveObject {
-  @HiveField(0)
-  late int bookingSerieIndex;
-
-  static void createGlobalState() async {
+class GlobalStateRepository extends GlobalStateInterface {
+  @override
+  void create() async {
     var globalBox = await Hive.openBox(globalStateBox);
     var startGlobalState = GlobalState()..bookingSerieIndex = 0;
     globalBox.add(startGlobalState);
   }
 
-  static Future<int> getBookingSerieIndex() async {
+  @override
+  Future<int> getBookingSerieIndex() async {
     var globalBox = await Hive.openBox(globalStateBox);
     GlobalState globalState = await globalBox.getAt(0);
     return globalState.bookingSerieIndex;
   }
 
-  static void increaseBookingSerieIndex() async {
+  @override
+  void increaseBookingSerieIndex() async {
     var globalBox = await Hive.openBox(globalStateBox);
     GlobalState globalState = await globalBox.getAt(0);
     globalState.bookingSerieIndex = globalState.bookingSerieIndex + 1;
     globalBox.putAt(0, globalState);
-  }
-}
-
-class GlobalStateAdapter extends TypeAdapter<GlobalState> {
-  @override
-  final typeId = globalStateTypeId;
-
-  @override
-  GlobalState read(BinaryReader reader) {
-    return GlobalState()..bookingSerieIndex = reader.read();
-  }
-
-  @override
-  void write(BinaryWriter writer, GlobalState obj) {
-    writer.write(obj.bookingSerieIndex);
   }
 }
